Fix MAIL TO option parsing for the case where the envelope-from has embedded space.
The current code order of parsing the options off the tail prior to
taking the address off the front seems bogus.  At best the option parser
has to deal with the address and discover it is not a known option.
When the address localpart has spaces (legal, if the localparte is quoted)
confusion ensues as the option-parser is not aware of the address quoting
rules.