Use xmlto; add to 'clean' target.
Instead of hardcoded xsltproc path and using zip ourselves, just use
xmlto, which delegates epub post-process to a script which handles all
this for us (doing the same thing).  Left previous commit in, so have
something to revert to if adding cover art or the like proves annoying
with xmlto in the way.

Added *.epub to "make clean"