Remove ACL-config skip so that ${acl } expansions work from queue-runs.
Previously we skipped parsing the ACL section when not needed.  Now it is
potentially needed in all cases.  The skip was ~5% faster than a full parse
so probably not a large part of the exim process startup.

Fix up testsuite output files affected by the removal and add a regression test.