Fix dovecot with empty 334 challenge.
Thomas Morper reported, with 4.82RC1, that he saw "334 NULL" as the
challenge when using AUTH PLAIN to Dovecot when the client does not send
an initial response.  I could replicate.

This was caused by commit 3f1df0e3 on 2012-11-19 (PP/13 of 4.82); I was
too cautious in the robustness fixes; the clue came in this line of
debug output:

    76430 dovecot: warning: ignoring trailing tab

This change removes that check, and documents in a comment that this
input is acceptable protocol-wise, and why.

With this fix:

    AUTH PLAIN
    334
    AGZyZWRlcmljAGh1bXB0eS1kdW1wdHk=
    235 Authentication succeeded