Unbreak HTML build for RC candidates.
The HTML build now uses the website repo, which extracts the version
number from the XML (generated from the .xfpt).  Meanwhile, commit
2aee48d6 made the version number in the build process dynamic, taking
the value from the release script (via an environ variable).

This change fixes the invocation to pass the version _without_ an RC
suffix to the XML build, letting HTML generation happen.