Finish the xfpt fix.
The readers of the version expect no surrounding whitespace; putting the
content in-place requires macro expansion, which requires that .literal
mode be `layout` or `off`, but either of those will escape the XML angle
brackets.  The easiest fix is to generate the XML as part of the macro
definition.

Moved local_params rules out of Makefile and into GenLocalParams