Add new ALLOW_SYSTEM_CRYPT_BRACES build option
FreeBSD now defaults to a {sha512} hash prefix.  Rather than continue
adding more types to crypteq{}{}, to have the same options available
everywhere, we allow this build option to cause the comparison to fall
through to the system crypt() for unrecognised {...} prefices -- the
same logic that happens if the crypt doesn't have a {...} prefix.

Set ALLOW_SYSTEM_CRYPT_BRACES in FreeBSD's OS Makefile.

The `expand.c` change was tested in a `#if 0` variant by Randy Bush, who
confirms that this works.  The rest is Exim build tuning and
documentation thereof, "too simple to go wrong".  *cough*