Point at readline fix on macOS
dlopen() without a qualified path searches env vars and paths stamped
into the binary; Exim is usually setuid, so this becomes just the paths
stamped into the binary.  On some platforms (*cough* macOS) libreadline
is not in the usual places.

So show how to stamp the Homebrew location of libreadline into the Exim
binary.