Add support for passing an open read-only file descriptor to clamd.
Additionally to the SCAN command clamd supports the zFILDES command to pass an
already open read-only file descriptor for the file to be scanned. The
advantage is that clamd does not need any special file system permissions to
scan files in exim's scan directory. The disadvantage is that clamd does not
know the file's name and will log a line like "fd[10]: OK".