Intercept chown()/fchown() failure and emit a pointer to the bugreport. Closes 2391
In a specific NFS setup we experienced a failing chown(). As it is not
clear, whether this was due to a misconfiguration or if this may happen in
other environments too, we behave as usual (abort the operation), but
issue a MAIN_LOG and PANIC_LOG entry pointing to this Bugreport.

You're encouraged to contact the developers, if you hit this issue.