Don't accept overly long IPv6 numbers
The current code stops parsing a hex number after 4 digits and then
loops around to look for punctuation.  If none is found, it just
starts parsing more hex digits, so it will accept invalid IPv6 such
as `1::1234567890abcdef`

My simple tweak makes it return 0 if a 5th digit is found.